+++ model/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "sequential-workflow-editor-model",
-	"version": "0.14.5",
+	"version": "0.14.6",
 	"homepage": "https://nocode-js.com/",
 	"author": {
 		"name": "NoCode JS",
