+++ model/src/value-models/choice/index.ts
@@ -1,1 +1,2 @@
+export * from './choice-value-model-configuration';
 export * from './choice-value-model';
